Adding back EnvConfigSource class and other changes
I realized that it's actually important to have the EnvConfigSource
object to enforce the type of configuration parameter resolution order
we want.

In this commit I also change the way I pass in ConfigSource objects to
the Context object. Looking at the old implementation I realize that it
is necessary to expose access to the raw data of the parsed
configuration files. I've mostly made these changes in reaction to
getting `context.channels` to behave like the old context object.